+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file(see http://k
 - map: write value to variable only when it exist.
 - ini: user can retrieve error position when syntax error.
 - json: user can retrieve error position when syntax error.
+- cmake: group files in the IDE automatically.
 
 ### Changed
 - fs: accept two types of lambda callbacks.